Refactor most page handlers into new pages/ directory. (#1226)
