+++ api/metricsdata_test.py
@@ -15,8 +15,8 @@ from __future__ import print_function
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import testing_config  # Must be imported first
 import unittest
-import testing_config  # Must be imported before the module under test.
 
 import datetime
 import mock
@@ -24,7 +24,7 @@ import flask
 
 from google.appengine.api import users
 
-import metrics
+from api import metricsdata
 import models
 
 
@@ -36,41 +36,41 @@ class MetricsFunctionTests(unittest.TestCase):
         bucket_id=1, property_name='prop')
 
   def test_truncate_day_percentage(self):
-    updated_datapoint = metrics._truncate_day_percentage(self.datapoint)
+    updated_datapoint = metricsdata._truncate_day_percentage(self.datapoint)
     self.assertEqual(0.01234568, updated_datapoint.day_percentage)
 
   def test_is_googler__anon(self):
     testing_config.sign_out()
     user = users.get_current_user()
-    self.assertFalse(metrics._is_googler(user))
+    self.assertFalse(metricsdata._is_googler(user))
 
   def test_is_googler__nongoogler(self):
     testing_config.sign_in('test@example.com', 111)
     user = users.get_current_user()
-    self.assertFalse(metrics._is_googler(user))
+    self.assertFalse(metricsdata._is_googler(user))
 
   def test_is_googler__googler(self):
     testing_config.sign_in('test@google.com', 111)
     user = users.get_current_user()
-    self.assertTrue(metrics._is_googler(user))
+    self.assertTrue(metricsdata._is_googler(user))
 
   def test_clean_data__no_op(self):
     testing_config.sign_in('test@google.com', 111)
     datapoints = [self.datapoint]
-    updated_datapoints = metrics._clean_data(datapoints)
+    updated_datapoints = metricsdata._clean_data(datapoints)
     self.assertEqual(0.0123456789, updated_datapoints[0].day_percentage)
 
   def test_clean_data__clean_datapoints(self):
     testing_config.sign_out()
     datapoints = [self.datapoint]
-    updated_datapoints = metrics._clean_data(datapoints)
+    updated_datapoints = metricsdata._clean_data(datapoints)
     self.assertEqual(0.01234568, updated_datapoints[0].day_percentage)
 
 
 class PopularityTimelineHandlerTests(unittest.TestCase):
 
   def setUp(self):
-    self.handler = metrics.PopularityTimelineHandler()
+    self.handler = metricsdata.PopularityTimelineHandler()
     self.datapoint = models.StableInstance(
         day_percentage=0.0123456789, date=datetime.date.today(),
         bucket_id=1, property_name='prop')
@@ -85,26 +85,26 @@ class PopularityTimelineHandlerTests(unittest.TestCase):
 
   def test_get_template_data__bad_bucket(self):
     url = '/data/timeline/csspopularity?bucket_id=not-a-number'
-    with metrics.app.test_request_context(url):
+    with metricsdata.app.test_request_context(url):
       actual = self.handler.get_template_data()
     self.assertEqual([], actual)
 
   def test_get_template_data__normal(self):
     testing_config.sign_out()
     url = '/data/timeline/csspopularity?bucket_id=1'
-    with metrics.app.test_request_context(url):
+    with metricsdata.app.test_request_context(url):
       actual_datapoints = self.handler.get_template_data()
     self.assertEqual(1, len(actual_datapoints))
     self.assertEqual(0.01234568, actual_datapoints[0]['day_percentage'])
 
 
-# TODO(jrobbins): Test for metrics.FeatureHandler.
+# TODO(jrobbins): Test for metricsdata.FeatureHandler.
 
 
 class FeatureBucketsHandlerTest(unittest.TestCase):
 
   def setUp(self):
-    self.handler = metrics.FeatureBucketsHandler()
+    self.handler = metricsdata.FeatureBucketsHandler()
     self.prop_1 = models.CssPropertyHistogram(
         bucket_id=1, property_name='b prop')
     self.prop_1.put()
@@ -125,14 +125,14 @@ class FeatureBucketsHandlerTest(unittest.TestCase):
     self.prop_4.delete()
 
   def test_get_template_data__css(self):
-    with metrics.app.test_request_context('/data/blink/cssprops'):
+    with metricsdata.app.test_request_context('/data/blink/cssprops'):
       actual_buckets = self.handler.get_template_data('cssprops')
     self.assertEqual(
         [(2, 'a prop'), (1, 'b prop')],
         actual_buckets)
 
   def test_get_template_data__js(self):
-    with metrics.app.test_request_context('/data/blink/features'):
+    with metricsdata.app.test_request_context('/data/blink/features'):
       actual_buckets = self.handler.get_template_data('features')
     self.assertEqual(
         [(4, 'a feat'), (3, 'b feat')],
