+++ pages/blink_handler.py
