+++ pages/users_test.py
@@ -15,14 +15,14 @@ from __future__ import print_function
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import testing_config  # Must be imported first
 import mock
 import unittest
-import testing_config  # Must be imported before the module under test.
 
 import flask
 
 import models
-import users
+from pages import users
 
 
 class SettingsHandlerTests(unittest.TestCase):
